requirements.txt
imgutils==0.1.2
Pillow==11.2.1
